Update API endpoints
Update API endpoints to use the correct endpoints for Staging and
Dev. They have been out of date since we moved from
"staging.testobject.org" to "app.staging.testobject.org". "branches" is
no longer used, so support for it is removed.

The UI is then built using `npm run build` to update the 'dist'
directory. This directory is what we are actually serving in product and
needs to be manually udpated.